requirements.txt
# Telegram Bot Framework
python-telegram-bot==21.9

# Environment Variables Management
python-dotenv==1.0.0

# Semantic Search
sentence-transformers>=2.3.0
torch>=2.0.0
numpy>=1.24.0
